add category and subcategory

+++ src/module/category/category.service.ts
@@ -9,9 +9,6 @@ export class CategoryService {
 
     async addCategory(payload: any, res: Response) {
         try {
-            let modifiedpaylod = {
-                name: payload.name
-            }
             //check it is already present...
             const iscategoryAlreadyPresent = await this.prisma.category.findFirst({
                 where: {
@@ -22,10 +19,13 @@ export class CategoryService {
                 throw new Error("this category is alredy exist...");
             }
             const data = await this.prisma.category.create({
-                data: payload.name
+                data: {
+                    name: payload.name
+                }
             })
             return res.status(HttpStatus.OK).json({ success: true, message: "cataegory created successfully.!", data });
         } catch (error) {
+            console.log(error);
             return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
         }
     }
@@ -61,10 +61,35 @@ export class CategoryService {
         try {
             const data = await this.prisma.category.findMany({
                 where: {
-                    parent_id: 0
+                    parent_id: null
                 }
             })
             return res.status(HttpStatus.OK).json({ success: true, message: "all cataegory fetch successfully.!", data });
+        } catch (error) {
+            return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
+        }
+    }
+    async getallsubcategories(res: Response) {
+        try {
+            const data = await this.prisma.category.findMany({
+                where: {
+                    NOT: {
+                        parent_id: null
+                    }
+                },
+                include: {
+                    parent: {
+                        select: {
+                            id: true,
+                            name: true
+                        }
+                    }
+                }
+
+            });
+            // console.log()
+
+            return res.status(HttpStatus.OK).json({ success: true, message: "all sub-cataegory fetch successfully.!", data });
         } catch (error) {
             return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
         }
@@ -81,12 +106,58 @@ export class CategoryService {
             return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
         }
     }
+    async getsubcategorysubcategory(id: number, res: Response) {
+        try {
+            const data = await this.prisma.category.findFirst({
+                select: {
+                    id: true,
+                    name: true,
+                    parent_id: true
+                },
+                where: {
+                    id: Number(id)
+                }
+            })
+            return res.status(HttpStatus.OK).json({ success: true, message: "subcategory fetch successfully.!", data });
+        } catch (error) {
+            return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
+        }
+    }
+    async sub_of_single_category(id: number, res: Response) {
+      
+        try {
+            const checkcategoriesexist = await this.prisma.category.findFirst({
+                where: {
+                    id: Number(id),
+                    parent_id: null
+                }
+            })
+            if (!checkcategoriesexist) {
+                return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: "category not found with this id" });
+            }
+            const data = await this.prisma.category.findMany({
+                where: {
+                    parent_id: Number(id)
+                },
+                include: {
+                    parent: {
+                        select: {
+                            name: true
+                        }
+                    }
+                }
+            })
+            return res.status(HttpStatus.OK).json({ success: true, message: "subcategory fetch successfully.!", data });
+
+        } catch (error) {
+            return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
+        }
+    }
     async editcategory(payload: any, id: number, res: Response) {
         try {
 
             //check category already present or not
-            const checkexist = await this.prisma.category.findFirst({ where: { id: Number(id) } })
-
+            const checkexist = await this.prisma.category.findFirst({ where: { id: Number(id), parent_id: null } })
             if (!checkexist) {
                 throw new Error("category with this is not found");
             }
@@ -108,14 +179,9 @@ export class CategoryService {
         try {
             //check already exist or not
             const checksubalreadyexist = await this.prisma.category.findFirst({
-                where: {
-                    AND: [
-                        {
-                            parent_id: Number(payload.parent_id)
-                        }, {
-                            name: payload.name
-                        }
-                    ]
+                where:
+                {
+                    id: Number(id)
                 }
             })
             if (!checksubalreadyexist) {
@@ -139,7 +205,7 @@ export class CategoryService {
         try {
 
             //check category already present or not..
-            const checkexist = await this.prisma.category.findFirst({ where: { id: Number(id) } })
+            const checkexist = await this.prisma.category.findFirst({ where: { id: Number(id), parent_id: null } })
             if (!checkexist) {
                 throw new Error("category with this is not found");
             }
@@ -157,16 +223,17 @@ export class CategoryService {
                     }
                 })
                 const ids = deleteIDS.map((ele) => ele.id);
-                //delete...
-                await this.prisma.category.deleteMany({
-                    where: {
-                        id: {
-                            in: ids
+                if (ids.length > 0) {
+                    //delete...
+                    await this.prisma.category.deleteMany({
+                        where: {
+                            id: {
+                                in: ids
+                            }
                         }
-                    }
-                })
+                    })
+                }
             }
-
             return res.status(HttpStatus.OK).json({ success: true, message: "cataegory deleted successfully.!", data });
         } catch (error) {
             return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
@@ -175,8 +242,7 @@ export class CategoryService {
 
     async deletesubCategory(id: number, res: Response) {
         try {
-
-            //check category already present or not..
+            //check subcategory already present or not..
             const checkexist = await this.prisma.category.findFirst({ where: { id: Number(id) } })
             if (!checkexist) {
                 throw new Error("sub-category with this is not found");
@@ -187,25 +253,7 @@ export class CategoryService {
                 }
 
             })
-            if (checkexist.parent_id === 0) {
-                const deleteIDS = await this.prisma.category.findMany({
-                    select: { id: true },
-                    where: {
-                        parent_id: data.id
-                    }
-                })
-                const ids = deleteIDS.map((ele) => ele.id);
-                //delete...
-                await this.prisma.category.deleteMany({
-                    where: {
-                        id: {
-                            in: ids
-                        }
-                    }
-                })
-            }
-
-            return res.status(HttpStatus.OK).json({ success: true, message: "cataegory deleted successfully.!", data });
+            return res.status(HttpStatus.OK).json({ success: true, message: "subcategory deleted successfully.!", data });
         } catch (error) {
             return res.status(HttpStatus.BAD_REQUEST).json({ success: false, message: error.message })
         }
